테스트 디렉토리 폴더 구조 변경
