--- src/main/resources/static/test.js
@@ -1,3 +1,0 @@
-/**
- * 
- */
