+++ src/test/java/com/hart/controller/LiveClassControllerTest.java
@@ -1,4 +1,4 @@
-package com.hart.ham.controller;
+package com.hart.controller;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
