+++ src/test/java/com/hart/mapper/RecipeMapperTest.java
@@ -1,4 +1,4 @@
-package com.hart.hoon;
+package com.hart.mapper;
 
 import java.util.List;
