+++ src/test/java/com/hart/service/LiveClassServiceTest.java
@@ -1,4 +1,4 @@
-package com.hart.ham.service;
+package com.hart.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
