+++ src/test/java/com/hart/service/ProductServiceTest.java
@@ -1,4 +1,4 @@
-package com.hart.hoon;
+package com.hart.service;
 
 
 import org.springframework.boot.test.context.SpringBootTest;
